requirements.txt
aniso8601==9.0.1
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.32.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector==2.2.9
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.1.1
